// src/app/core/models/requiredDocumentDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface RequiredDocumentDto {
  id: string;
  tenantId: string;
  documentName: string;
  description: string;
  documentType: string;
  entityType: string;
  isRequired: boolean;
  isActive: boolean;
  allowedFileTypes: string;
  maxFileSizeBytes: number;
  createdBy: string;
  updatedBy: string;
  createdAt: string;
  updatedAt: string;
}